+++ packages/grid/src/features/standard-table/features/column-groups/ColumnGroupRow.tsx
@@ -1,5 +1,5 @@
 import { Indent, Row } from "@stenajs-webui/core";
-import { ZIndexProperty } from "csstype";
+import { Property } from "csstype";
 import * as React from "react";
 import {
   defaultTableRowHeight,
@@ -10,6 +10,7 @@ import { TableHeadRow } from "../../../table-ui/components/table/TableHeadRow";
 import { useGroupConfigsForRows } from "../../context/GroupConfigsForRowsContext";
 import { useStandardTableConfig } from "../../hooks/UseStandardTableConfig";
 import { ColumnGroupItem } from "./ColumnGroupItem";
+import ZIndex = Property.ZIndex;
 
 interface ColumnGroupRowProps {
   height?: string;
@@ -37,7 +38,7 @@ export const ColumnGroupRow = React.memo(function ColumnGroupRow({
       shadow={stickyHeader ? "var(--swui-sticky-header-shadow)" : undefined}
       zIndex={
         stickyHeader
-          ? zIndex ?? ("var(--swui-sticky-header-z-index)" as ZIndexProperty)
+          ? zIndex ?? ("var(--swui-sticky-header-z-index)" as ZIndex)
           : zIndex
       }
     >